pyproject.toml
[tool.poetry]
name = "banya-rest"
version = "0.1.0"
description = ""
authors = ["ksm <ksm@daiblab.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
django = "^5.1.2"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.10"
numpy = "^2.1.3"
openai = "^1.53.0"
transformers = "^4.46.1"
pytz = "^2024.2"
google = "^3.0.0"
google-cloud-storage = "^2.18.2"
cryptography = "^43.0.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
